requirements.txt
absl-py==2.1.0
anyio==3.7.1
astunparse==1.6.3
audioread==3.0.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
efficientnet==1.0.0
exceptiongroup==1.2.0
ffmpeg-python==0.2.0
Flask==1.1.4
flatbuffers==24.3.7
fonttools==4.51.0
future==1.0.0
gast==0.4.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
h11==0.12.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.6
image-classifiers==1.0.0
imageio==2.34.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.4.0
keras==2.15.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==1.1.1
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.8
namex==0.0.7
networkx==3.2.1
norbert==0.2.1
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==1.5.3
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.12.0
segmentation-models==1.0.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.2.12
typer==0.3.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==1.0.1
wrapt==1.14.1
